PHO-29: Changed Receipt State enum and renamed launch handler method
